% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bittner_railway}
\alias{bittner_railway}
\title{Bittner Creek railway}
\format{
An \code{sf} data frame with columns:
\describe{
\item{track_classification}{Railway track classification}
\item{geom}{Linestring geometry (WGS 84)}
}
}
\source{
BC GBA Railway Tracks via newgraph database
}
\usage{
bittner_railway
}
\description{
Railway track segments within expanded Bittner Creek watershed.
}
\keyword{datasets}
