% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bittner_roads}
\alias{bittner_roads}
\title{Bittner Creek roads}
\format{
An \code{sf} data frame with columns:
\describe{
\item{road_type}{Transport line type code (RH1, RA1, RA2, etc.)}
\item{geom}{Linestring geometry (WGS 84)}
}
}
\source{
BC Digital Road Atlas via newgraph database
}
\usage{
bittner_roads
}
\description{
DRA road segments within expanded Bittner Creek watershed.
}
\keyword{datasets}
