% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neexdzii_bc}
\alias{neexdzii_bc}
\title{British Columbia boundary (simplified)}
\format{
An \code{sf} data frame with 1 row and polygon geometry (WGS 84).
}
\source{
BC Freshwater Atlas via newgraph database
}
\usage{
neexdzii_bc
}
\description{
Simplified BC provincial boundary for keymap inset context.
}
\keyword{datasets}
