% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neexdzii_falls}
\alias{neexdzii_falls}
\title{Neexdzii Kwa falls}
\format{
An \code{sf} data frame with point geometry (WGS 84).
}
\source{
bcfishpass via newgraph database and fresh package
}
\usage{
neexdzii_falls
}
\description{
Known falls/barriers from bcfishpass within the Neexdzii Kwa subbasin.
}
\keyword{datasets}
