% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neexdzii_fish_obs}
\alias{neexdzii_fish_obs}
\title{Neexdzii Kwa fish observations}
\format{
An \code{sf} data frame with columns including:
\describe{
\item{species_code}{Species code (CO, CH, RB, etc.)}
\item{geom}{Point geometry (WGS 84)}
}
}
\source{
bcfishobs via newgraph database and fresh package
}
\usage{
neexdzii_fish_obs
}
\description{
Fish observation events from FISS within the Neexdzii Kwa subbasin.
}
\keyword{datasets}
