% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neexdzii_lakes}
\alias{neexdzii_lakes}
\title{Neexdzii Kwa lakes}
\format{
An \code{sf} data frame with polygon geometry (WGS 84).
}
\source{
BC Freshwater Atlas via newgraph database and fresh package
}
\usage{
neexdzii_lakes
}
\description{
FWA lake polygons matched by waterbody key on the stream network
within the Neexdzii Kwa subbasin.
}
\keyword{datasets}
