% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neexdzii_railway}
\alias{neexdzii_railway}
\title{Neexdzii Kwa railway}
\format{
An \code{sf} data frame with columns:
\describe{
\item{track_name}{Railway track name}
\item{geom}{Linestring geometry (WGS 84)}
}
}
\source{
BC GBA Railway Tracks via newgraph database
}
\usage{
neexdzii_railway
}
\description{
Railway track segments clipped to the Neexdzii Kwa subbasin.
}
\keyword{datasets}
